Develop

+++ app/src/main/java/com/smxy/hencoder/util/Util.kt
@@ -1,13 +1,12 @@
 package com.smxy.hencoder.util
 
-import android.content.Context
 import com.smxy.hencoder.MyApplication
 
 class Util {
     companion object {
-        fun dip2px(dpValue: Float): Int {
+        fun dip2px(dpValue: Float): Float {
             val scale = MyApplication.context.getResources().getDisplayMetrics().density
-            return (dpValue * scale + 0.5f).toInt()
+            return dpValue * scale + 0.5f
         }
     }
 }
